// frontend/src/types.ts
export interface CustomFileBase extends Partial<globalThis.File> {
  processingTotalTime: number | string;
  status: string;
  nodesCount: number;
  relationshipsCount: number;
  model: string;
  fileSource: string;
  sourceUrl?: string;
  wikiQuery?: string;
  gcsBucket?: string;
  gcsBucketFolder?: string;
  errorMessage?: string;
  uploadProgress?: number;
  processingStatus?: boolean;
  googleProjectId?: string;
  language?: string;
  processingProgress?: number;
  accessToken?: string;
  isChecked?: boolean;
  retryOptionStatus: boolean;
  retryOption: string;
  chunkNodeCount: number;
  chunkRelCount: number;
  entityNodeCount: number;
  entityEntityRelCount: number;
  communityNodeCount: number;
  communityRelCount: number;
  createdAt?: Date;
  token_usage: number;
  embedding_model?: string;
}
// ...
export interface S3ModalProps {
  hideModal: () => void;
  open: boolean;
}
export interface GCSModalProps extends Omit<S3ModalProps, ''> {
  openGCSModal: () => void;
}
// ...
export interface fileName {
  fileName: string;
  fileSize: number;
  url: string;
  gcsBucketName?: string;
  gcsBucketFolder?: string;
  status?: string;
  gcsProjectId: string;
  language?: string;
}
// ...
export interface nonoautherror extends NonOAuthError {
  message?: string;
}